telegram-bot/requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
structlog==23.2.0
